Create own file for LegacyRemoteTabsTableViewController
